Re-arrangd components diretory: put player-controls in a separate folder
