+++ src/components/common/player/components/player-controls/PlayerControls.js
